+++ problem-test/project-car-mgt-api/application/src/main/kotlin/dudgns/com/backend/carmgt/application/dto/carManagement/GetCarInfoListQueryCommand.kt
@@ -1,9 +1,9 @@
 package dudgns.com.backend.carmgt.application.dto.carManagement
 
 data class GetCarInfoListQueryCommand (
-    override var companyCode: Int?,
+    override var companyCode: Long?,
     override var rentalYn: Boolean?,
     override var startYear: Int?,
     override var endYear: Int?,
-    override var categoryTypeList: List<Long?>?
+    override var categoryTypeList: List<Int?>?
 ) : IGetCarInfoListDto
